+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # v4.1.2
-## mm/dd/2019
+## 06/22/2019
 
 1. [](#new)
   * Added new `h#` tags for `h1` through `h6` supporting `class` and `id` attributes