+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# v4.1.2
+## 06/22/2019
+
+1. [](#new)
+  * Added new `h#` tags for `h1` through `h6` supporting `class` and `id` attributes
+1. [](#improved)
+  * Make `ShortcodeManager::setStates()` more flexible to accept any type of object  
+
 # v4.1.1
 ## 04/23/2019